requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
async-generator==1.10
attrs==22.1.0
certifi==2022.12.7
cffi==1.15.1
colorama==0.4.6
exceptiongroup==1.0.4
h11==0.14.0
idna==3.4
iniconfig==1.1.1
Mako==1.2.4
MarkupSafe==2.1.1
outcome==1.2.0
packaging==22.0
parse==1.19.0
parse-type==0.6.0
pluggy==1.0.0
psutil==5.9.4
py==1.11.0
pycparser==2.21
PySocks==1.7.1
pytest==7.2.0
pytest-bdd==6.1.1
pytest-html==3.2.0
pytest-metadata==2.0.4
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.13
wsproto==1.2.0
